Addressed the peoject invalid state when groups rejects to be on the project and it's manager is the project lead.
1)  Implemented the functionality to redirect the group manager who is rejecting to be  on the the project to select a new lead before they can reject the invitation. This makes prevents the project invalid state.

2) Fixed compiler error caused by adding new fields to controllers that wasn't addressed in the tests classes for Gorup controller, and Home controller.
