+++ code/TicketApp/TicketAppWeb/Controllers/HomeController.cs
@@ -10,10 +10,12 @@ namespace TicketAppWeb.Controllers;
 public class HomeController : Controller
 {
     private readonly IProjectRepository _projectRepository;
+    private readonly IGroupRepository _groupRepository;
 
-    public HomeController(IProjectRepository projectRepository)
+    public HomeController(IProjectRepository projectRepository, IGroupRepository groupRepository)
     {
         _projectRepository = projectRepository;
+        _groupRepository = groupRepository;
     }
 
     /// <summary>
@@ -43,28 +45,43 @@ public class HomeController : Controller
         try
         {
             await _projectRepository.ApproveGroupForProjectAsync(projectId, groupId);
-            return Json(new { success = true });
+            return RedirectToAction("Index", "Home");
         }
         catch (Exception ex)
         {
-            return Json(new { success = false, message = ex.Message });
+            TempData["ErrorMessage"] = "Error processing approval: " + ex.Message;
+            return RedirectToAction("Index", "Home");
         }
     }
 
+
     /// <summary>
-    /// Rejects a group for a project.
+    /// Rejects a group for a project. If the group manager is set lead of project, 
+    /// redirects to edit the project to set a new lead
     /// </summary>
     [HttpPost]
     public async Task<IActionResult> RejectGroupForProject(string projectId, string groupId)
     {
+        var project = await _projectRepository.GetProjectByIdAsync(projectId);
+        if (project == null) return RedirectToAction("Index", "Home");
+
+        var group = await _groupRepository.GetAsync(groupId);
+        if (group == null) return RedirectToAction("Index", "Home");
+
+        if (project.LeadId == group.ManagerId)
+        {
+            return RedirectToAction("EditProject", "Project", new { id = projectId, leadChangeRequired = true });
+        }
+
         try
         {
             await _projectRepository.RejectGroupForProjectAsync(projectId, groupId);
-            return Json(new { success = true });
+            return RedirectToAction("Index", "Home");
         }
         catch (Exception ex)
         {
-            return Json(new { success = false, message = ex.Message });
+            TempData["ErrorMessage"] = "Error processing request: " + ex.Message;
+            return RedirectToAction("Index", "Home");
         }
     }
