+++ code/TicketApp/TicketAppWeb/Models/DataLayer/Repositories/ProjectRepository.cs
@@ -404,6 +404,11 @@ public class ProjectRepository(TicketAppContext ctx) : Repository<Project>(ctx),
             .ToListAsync();
     }
 
+    /// <summary>
+    /// Gets the projects by lead asynchronous.
+    /// </summary>
+    /// <param name="leadId">The lead identifier.</param>
+    /// <returns></returns>
     public async Task<List<Project>> GetProjectsByLeadAsync(string leadId)
     {
         return await context.Projects.Where(p => p.LeadId == leadId).ToListAsync();
