+++ code/TicketApp/TicketAppWeb/Views/Project/AddProject.cshtml
@@ -171,5 +171,3 @@
         });
     </script>
 }
-
-}
